Moved clone_subset to split
