+++ src/mesh/merging_and_splitting.rs
@@ -69,51 +69,6 @@ impl Mesh
         self.create_boundary_edges();
     }
 
-    /// Returns a clone a subset of this mesh.
-    pub fn clone_subset(&self, faces: &std::collections::HashSet<FaceID>) -> Mesh
-    {
-        let info = crate::mesh::ConnectivityInfo::new(faces.len(), faces.len());
-        for face_id in faces {
-            let face = self.connectivity_info.face(*face_id).unwrap();
-            for halfedge_id in self.face_halfedge_iter(*face_id) {
-                let mut walker = self.walker_from_halfedge(halfedge_id);
-                let halfedge = self.connectivity_info.halfedge(halfedge_id).unwrap();
-                info.add_halfedge(halfedge_id, halfedge);
-
-                let vertex_id = walker.vertex_id().unwrap();
-                let vertex = self.connectivity_info.vertex(vertex_id).unwrap();
-                info.add_vertex(vertex_id, vertex);
-                info.set_vertex_halfedge(vertex_id, walker.next_id());
-
-                walker.as_twin();
-                if walker.face_id().is_none()
-                {
-                    let twin_id = walker.halfedge_id().unwrap();
-                    let twin = self.connectivity_info.halfedge(twin_id).unwrap();
-                    info.add_halfedge(twin_id, twin);
-
-                }
-                else if !faces.contains(&walker.face_id().unwrap())
-                {
-                    let twin_id = walker.halfedge_id().unwrap();
-                    let mut twin = self.connectivity_info.halfedge(twin_id).unwrap();
-                    twin.face = None;
-                    twin.next = None;
-                    info.add_halfedge(twin_id, twin);
-                }
-            }
-
-            info.add_face(*face_id, face);
-        }
-
-        let mut positions = HashMap::with_capacity(info.no_vertices());
-        for vertex_id in info.vertex_iterator() {
-            positions.insert(vertex_id, self.vertex_position(vertex_id).clone());
-        }
-
-        Mesh::new_internal(positions, info)
-    }
-
     ///
     /// Merges overlapping faces, edges and vertices.
     ///
